pyproject.toml
[project]
name = "hr-agent"
version = "2.0.0"
description = "An enterprise-grade HR assistant built with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "ACME Corporation"}
]
keywords = ["langchain", "langgraph", "langfuse", "hr", "agent", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.40.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.30",
    "rich>=13.7.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.31.0",
    "pyyaml>=6.0.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langfuse>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hr_agent", "apps"]

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hr_agent"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
